requirements.txt
Django==1.7.3
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
Sphinx==1.2.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
bpython==0.13.1
certifi==14.05.14
chardet==2.1.1
django-allauth==0.19.0
django-bootstrap3==5.1.1
django-contrib-comments==1.5
django-crispy-forms==1.4.0
django-csvimport==2.4
django-debug-toolbar==1.2.2
django-debug-toolbar-template-profiler==1.0.1
django-debug-toolbar-template-timings==0.6.4
django-debugtools==1.2.1
django-extensions==1.4.9
django-mptt==0.6.1
django-mptt-admin==0.2.0
docutils==0.12
ipython==2.3.1
json-table-schema==0.1
lxml==3.4.1
matplotlib==1.4.2
messytables==0.14.1
mock==1.0.1
nose==1.3.4
numpy==1.9.1
numpydoc==0.5
oauthlib==0.7.2
pandas==0.15.2
pyparsing==2.0.3
python-dateutil==1.5
python-magic==0.4.6
python-openid==2.2.5
pytz==2014.10
pyzmq==14.5.0
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
sqlparse==0.1.14
tornado==4.0.2
wsgiref==0.1.2
xlrd==0.9.3
